Added R/W property NSString *text to KDocument which allows for high-level manipulation of the document's content. Also made line range machinery thread safe by employing a light-weight spinlock. The edited timestamp of a KDocument is now also thread safe and stored as a 64-bit unsigned integer represeting a microsecond-resolution timestamp rather than a double-precision floating point representing seconds.
**Discussion about the line range spin lock:**

Since the line range machinery is called on the main thread in almost all cases, there should be a very low probability for contention, thus a super spin lock is a good choise. If we perform more line range calculations in other threads in the future, we might consider re-implementing this using detailed CAS operations instead.

**resources/main.js:**

Comment-out the "tabDidBecomeSelected" log message since it caused an aweful lot of output. Also, added a (commented-out) setter, demonstrating how the node API can be used to replace or manipulate documents (in this case by replacing the contents)